chore: change files in color tailwind
